Categorize commands WIP
