+++ source/game_sa/Scripts/Commands/CLEO/Char.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+/*!
+* Various utility commands
+*/
+
+template<>
+OpcodeResult CRunningScript::ProcessCommand<COMMAND_STORE_CLOSEST_ENTITIES>() { // 0x0AB5 (Char self)
+    CollectParameters(1);
+    return OR_CONTINUE;
+}
+
