+++ source/game_sa/Scripts/Commands/CLEO/Game.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+/*!
+* Various CLEO game commands
+*/
+
+template<>
+OpcodeResult CRunningScript::ProcessCommand<COMMAND_IS_GAME_VERSION_ORIGINAL>() { // 0x0AA9
+    CollectParameters(0);
+    return OR_CONTINUE;
+}
+
